enum to separate folder, getResource in test
