+++ src/main/java/com/example/test/bean/GroupedProduct.java
@@ -5,6 +5,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import com.example.test.enums.Currency;
 import com.google.gson.annotations.SerializedName;
 
 import lombok.Data;
