+++ src/test/java/XHTMLParserTest.java
@@ -1,5 +1,3 @@
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.List;
 
 import org.junit.Test;
@@ -12,17 +10,17 @@ import org.junit.Before;
 
 public class XHTMLParserTest {
 
-    private Path testFilePath;
+    private String testFilePath;
 
     @Before
     public void init() {
-        this.testFilePath = Paths.get("src\\test\\java\\testFiles\\inputExample.xhtml");
+        testFilePath = getClass().getClassLoader().getResource("inputExample.xhtml").getFile();
     }
 
 
     @Test
     public void testXHTMLParser(){
-        List<Product> products = XHTMLParser.parse(this.testFilePath.toString());
+        List<Product> products = XHTMLParser.parse(this.testFilePath);
         Assert.assertEquals(7, products.size());
         Assert.assertEquals(1, (int) products.get(0).getId());
         Assert.assertEquals("Product 1", products.get(0).getName());
