+++ src/test/resources/inputExample.xhtml
